Добавлена кнопка "Удалить пост". Исправлена ошибка, связанная с неправильной информацией о блоге, в котором создан пост

+++ blogsite/views.py
@@ -1,4 +1,5 @@
-from django.shortcuts import render
+from django.shortcuts import render, redirect
+from django.core.urlresolvers import reverse
 from django.views import View
 from django.http import HttpResponse
 from blogsite.models import Post, Blog, BlogSubscriber, UserPostWatched
@@ -35,7 +36,7 @@ class NewPost(View):
             post.post_text = form['post_text'].value()
             post.blog_id = get_blog_id(request.user.id)
             post.save()
-            return HttpResponse('Created')
+            return redirect(reverse('blogsite:posts', kwargs={'blog_id': post.blog_id}))
         else:
             return HttpResponse('Not created!')
 
@@ -44,9 +45,12 @@ class DeletePost(View):
     def get(self, request, post_id):
         post = Post.objects.filter(id=post_id)
         if post:
+            blog = Blog.objects.get(id=post.values()[0]['blog_id'])
+            if request.user.id != blog.user_id:
+                return redirect(reverse('blogsite:feed'))
             post.delete()
-            return HttpResponse('Deleted')
-        return HttpResponse('Not deleted')
+            return redirect(reverse('blogsite:posts', kwargs={'blog_id': blog.id}))
+        return redirect(reverse('blogsite:feed'))
 
     def post(self, request):
         pass
@@ -56,8 +60,8 @@ class ShowPost(View):
     def get(self, request, post_id):
         post = Post.objects.get(id=post_id)
         post.created_at = post.created_at.strftime("%d.%m.%Y %H:%M")
-        blog = Blog.objects.get(id=get_blog_id(request.user.id))
-        args = {'post': post, 'blog': blog}
+        blog = Blog.objects.get(id=post.blog_id)
+        args = {'post': post, 'blog': blog, 'isMyPost': request.user.id==blog.user_id}
         return render(request, 'blogsite/post.html', args)
 
     def post(self, request):
@@ -98,7 +102,6 @@ class MarkAsRead(View):
 class NewsFeed(View):
 
     def get(self, request):
-        print (request.user.is_authenticated())
         my_subcriptions = BlogSubscriber.objects.filter(user_id=request.user.id)
         blog_ids = [x['blog_id'] for x in my_subcriptions.values()]
         my_feed = Post.objects.filter(blog_id__in=blog_ids).values()
